backend/requirements.txt
fastapi
uvicorn
pydantic
requests
apscheduler
python-dotenv
sqlalchemy
reportlab
